Rename package to tc_core.
