+++ setup.py
@@ -4,13 +4,13 @@ from setuptools import setup
 
 
 setup(
-    name='thumbor_community',
+    name='tc_core',
     version='0.1',
-    url='http://github.com/thumbor_community/thumbor_community',
+    url='http://github.com/thumbor_community/core',
     license='MIT',
     author='Thumbor Community',
     description='Thumbor community extensions core',
-    packages=['thumbor_community'],
+    packages=['tc_core'],
     include_package_data=True,
     zip_safe=False,
     platforms='any',
