+++ tc_core/context_importer.py
@@ -2,7 +2,7 @@
 
 from thumbor.context import ContextImporter as ThumborContextImporter
 from thumbor.utils import logger
-from thumbor_community.importer import Importer
+from tc_core.importer import Importer
 
 
 class ContextImporter(ThumborContextImporter):
